requirements.txt
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
black==23.1.0
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
dj-database-url==1.2.0
Django==4.1.7
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.0
executing==1.2.0
Flask==2.2.3
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
ipdb==0.13.11
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
MouseInfo==0.1.3
mypy-extensions==1.0.0
packaging==23.0
parso==0.8.3
pathspec==0.11.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pure-eval==0.2.2
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
Pygments==2.14.0
PyJWT==2.6.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
python-dotenv==1.0.0
pytweening==1.0.4
pytz==2022.7.1
pywhatkit==5.4
PyYAML==6.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.3
whitenoise==6.4.0
wikipedia==1.4.0
